Create script for loading env variables
This way, dotenv filepath only needs to be configured once
on the server (was running into issues with filepath of .env
when loading the server on its own or from the root package
with npm run dev), and it can give useful errors if an env
variable has not been set.

Previously when I had forgotten to set an env variable, the
errors thrown during deployment tended to make the source
of the problem less obvious.