Move prisma from server sub package to root package
This was done since the env path cannot be changed
on prisma. The project has to look for the env file
in the root directory otherwise deploy won't work.

Therefore prisma has been moved out of server package
and into the root package as a dependency, so prisma
can access the root dir's env file. It is not a
dependency of the client project.

How very annoying and stupid that you can't customise
the env path on prisma. How incredibly annoying.