Unify tests with vitest on the root package
I did not realise vitest could be used as a drop-in
replacement for jest. Now instead of separate tests
ran on client and server (with vitest and jest, re-
spectively), the tests are run on the root package
with vitest. This also stops having to use weird
command line settings for node experimental vm mo-
dules. Vitest just works.

There is however an error from one of the tests in
api.v1.test. Think the separate dev and prod routes
are not being routed to the intended controller
functions. Need to rethink 'isDevMode' in favour of
a 'isAdminMode' middleware - using dev mode env to
mean 'I want to make admin changes and upload ima-
ges' is causing all kinds of issues. So change that,
and then fix the tests. Plz!