Client fix: broke npm dev after removing VITE_SERVER_URL from fetch urls
On docker tests and production, can omit VITE_SERVER_URL from fetch
since client is grabbed from server, everything is on the same port.
However when running npm run dev, the client currently runs on 5173
and the server on 8080.

So instead of omitting the VITE_SERVER_URL from the fetch urls, just
use a blank string for its value in the env variables.

Had to change server/src/env checks since apparently an empty string
is close enough to undefined to evaluate as such, or at least the way
dotenv was loading it, it ended up evaluating !VITE_SERVER_URL to true.