Fix: integration tests not running in container
Had some strange stuff going on with tests - then
realised that the docker containers were being started,
then vitest was being run on the host machine, not the
containerized application.

So I created a Dockerfile.dev which is used in the
docker-compose file's server build. It merely changes
the start up command to run vitest instead of npm run
start - I tried to get this working with a command in
the docker-compose.yml file, but it just didn't seem
to work.

Docker compose is also set up to watch and sync any
changes in the local filesystem, so when a test or
file changes, the file within the container is updated,
vitest running in the container detects the change,
and the tests re-run.