Fix: db imageTag posX, posY stored as decimal
This was causing problems when comparing to js numbers,
as decimal is not an inbuilt js type, so it seems to
basically be treated as a string. Made comparing db
entries in tests to numeric values very annoying.

So now prisma stores them as floats (which when
grabbed from the db, are js type number) instead
of decimal.