integration tests: can now see express errors during tests
Usually the main app would catch errors and deal with
them, printing to console and sending error to client or
whatever it is.

However since test is running outside of normal express
app, any errors thrown by controllers would just end up
disappearing. So this middleware just logs the error
to console before the test fails, which makes it a lot
easier to debug everything.