Fix: exports from env.mjs were constant upon load and therefore immutable
Once the env vars were loaded, they were exported from the mjs file.
However, I am in the process of turning admin functionality from
hardcoded in the env file to something that can be turned on or off
if the user enters the correct password. So the env variable needed
to be mutable. However, using an exported const from env.mjs was not
mutable - so I have switched back to using process.env to access env
vars.

However, env.mjs will still be used to load and check for the existence
of all the correct env vars for the application to run. This provides
useful feedback, e.g. if you deploy the application but have forgotten
to set up all the proper env vars. Without this feedback, typically you
get cryptic errors from where the application tries to run without the
correct env vars.

To do: export a function from env.mjs to load the env vars. At the moment,
it is just a line in the file that loads the env vars, but it's not clear
when or why this code actually runs. If it is an exportable function I
will also be able to use it in tests as well to make sure the correct env
vars are loaded.