Re-organise integration tests to improve efficacy
Previously, I was testing separate components of the
application, i.e. each router separately. This was
kind of missing the point of integration tests, and
I found myself having to fake a bunch of stuff -
creating an express app, add the router being tested,
set up the storage volume, etc. etc. and I realised,
wow this is stupid and misses the point of integration
tests.

So I have gotten rid of all that, and now the tests
within tests/ all work with the app exported from
app.mjs. However the tests are split up into multiple
files for organisation. Think I have finally got a
good handle on how to structure the tests for each
API endpoint and method.

And since I have done a better implementation of
enabling/disabling admin mode in a previous commit,
those can now be tested together easily.