client: rename components to reflect their usage
EditModePhotoTagOverlay and EditModePhoto can be used
for view mode purposes, so renaming since they will be
generic components used in both modes, not specific
just to one mode.

Likewise, PhotoDetails is for edit mode only so it
has been renamed to reflect that.

Good job making EditModePhoto general enough that it
could just be dropped into ViewMode and with the provided
props, do what is needed!