client: ViewModePage: avoid repeating same image twice in a row
This was surprisingly a pain to do. It turns out,
the values provided to each callback are a snapshot,
and immutable. I expected the callback functions to
be re-defined when state changed, but this seems to
not be the case, as callbacks would end up using v-
alues from a previous render.

React refs are used to keep a reference to the latest
selected photo id, and the completed ids.

Might have been able to use useCallback hook to get
the functions to be re-defined upon their dependencies
changing. I'm sure the callbacks get re-defined upon
re-render, but it seems like they aren't getting re-
defined with the new state values. Very confusing.

Apparently this problem is called "stale closures".