client: create useDarkThemeActive hook
This will make it easy to show or modify components
based on whether the user has dark mode on or not.
I was going to do that in CSS, but SVG icons don't
seem to work when I try to import them and link as
a URL in styled-components. So the reason for this
hook is to change the icon when in dark/light mode.