Replace supertest with sagetest
Getting strange errors when implementing sessions,
due to a bug in supertest or one of its dependencies.

Detailed here:

https://github.com/forwardemail/supertest/issues/830

Had to rework some of the tests a little with how
form fields are given to the request (supertest was
also supposd to be able to take js objects, but this
never worked for me, so I ended up using strings.
Another reason to avoid it in future), but everything
is working now.