server: make temp fix for railway deployment not setting cookie
When cookie is secure, it does not get set on the
deployment on railway, despite the fact that the
server serves up the client, so the domains are
the same, and everything is done over https.

When secure is set to false, it sets the cookie
just fine. I have done much searching and can find
nothing for this particular problem, so I will
simply set secure to false for now and try to get
on with some other things instead of continuing
to bang my head against a wall.