server > /api/v1/check-tag: return all found tags instead of just the newly found tag
This goes against the idea of the api being stateless.
You should get the same output with the same input.

However in this instance, the server is keeping track
of all the tags found by the user, and the client is
having to keep track of it too, since check-tag just
returns the newly found tag. The client has to keep
an array of found tags and update it, etc.

Since the server has to keep track of this anyway,
in order to know when the player has 'won', I may
as well just have the check-tag route return all
found tags so the client doesn't have to keep track
of it separately. Hopefully this doesn't turn out
to be a bad idea...