client > styled > FakeLink: create component to style links that go nowhere and supress default behaviour
Default links move the screen to the top when clicked,
and we want to prevent this if we are just clicking
a link on a single page application. But links also
work with keyboard navigation by default so are good
for accessibility.

Using a span instead of a link and styling that to
look like a link means we lose the keyboard navigation
and would have to re-implement ourselves, i.e., a
waste of time.