fix: client > ViewWinModal: onClose method getting called twice when dialog closed via button
This was because the callback set modal active to false,
then called the api. Because the modal had the same callback
for onClose property, it would then call it again. The
modal component calls onClose based on DOM events with
a ref, i.e. whenever isActive gets changed, the effect
runs and checks if it was closed, and if so, invokes
the onClosed callback.