server > /api/v1: save foundAllTags to session
The post check-tag route already checks for whether
the user has found all tags so a message can be sent
to the client saying so. The result is now saved in
the session data, so that the post time route doesn't
have to make this check again to make sure a user has
really found all the tags on an image before posting
a time to the database.

Reducing duplication, basically.