client > shared.types > Tag: simplify with utility types
Since tags not commited to server can be created on client,
and once given to server will have ids created for them,
i.e. on tag editing components there will be a mix of tags
with and without ids. Therefore, id and imageId have been
made optional, and the Required<> utility type can be used
in situations where the optional id and imageId are needed,
i.e. we are dealing with server tags only.