client > ext > simplifyServerValidationErrors: make output better suited to providing form errors
Output is now an object instead of an array. It has
an array property for easily iterating through all
the errors, but there are also properties on the
output object. Each key corresponds to the name of
the form field. This makes it easy to grab each error
for each of the form inputs.

Each key corresponds to an array of errors, since
validator can respond with multiple errors for the
same field (i.e. empty, not long enough, etc.).
To get just one error, access like output[key][0].