client > ViewModePhoto: make PhotoWithTagOverlays more flexible
So its parent can handle what the contents of the
currently selected tag is. This may differ between
view and edit modes, so better to have it handled
by the parent than hardcoded in the component.

E.g. in view mode the name might be '???' and in
edit mode, 'new tag', or 'currently selected', etc.