client > Form buttons: stop forced justify-end in certain situations
CSS priority was given to form child button element
with media query, as it was more specific, so impossible
to override in a logical way, so I just got rid of it.

Originally, the buttons would justify-end when the screen
width was above 500px. However on the win modal I want
the button to be centered. This was impossible due to
the CSS specifity.

Having different buttons or mixins or something might
have been good. Also putting the breakpoints in a
file so they could be imported from one place and used
throughout all styled-components.