client > ViewWinTimeFormFields: add client validation
This will not listen until the input field has been
blurred (lost focus). Then it will report errors for
the first time, and react and provide updated errors
based on the user's input. Once the value is valid,
errors won't be checked again until the input field
has lost focus again.

This is to avoid the input field immediately showing
an error before the user has even touched the field,
or before they have finished typing; both of which I
have always found to be annoying.