client: create generic ValidatedInput component
This will check validity of the input once blur has
occurred, but not before. Then on input, the errors
will be updated until the input value is valid, at
which point errors won't be checked or shown to the
user until they blur out of the input once more.

Basically it makes errors way less annoying. Nothing
worse than a form moaning at you when you are just
looking at it for the first time and haven't even
typed anything, or haven't finished typing in a field.