client > PhotoTagsForm: fix ValidatedInput making changing posX/posY a pain
Moving the tag would produce a very long decimal number
between 0 and 1, and since the step was 0.01 or something,
it would require the value to be within one of those steps.
So e.g. something like 0.015 or anything more precise than
that would fail validation.

Don't like the way posX/posY work now - if you try
typing any non-number characters it just doesn't let
you, and if you type a number it doesn't like, it
doesn't just truncate it but remove it completely.
Pretty annoying.