client > EditPhotoDetails: fix form not filling up grid column
This was due to the margin: 0 auto on the base container styled
component. Now PaddedContainer literally just has padding,
and I will rename Container to CenteredContainer to make it
clearer what it does (and it has no padding).

When centered and padding is required, one of each can
be used, nested.